add settings to the graphical interface

+++ HospitalEmergencySimulation/DataForm.xaml.cs
@@ -15,9 +15,6 @@ using System.Windows.Shapes;
 
 namespace HospitalEmergencySimulation
 {
-    /// <summary>
-    /// Lógica de interacción para DataForm.xaml
-    /// </summary>
     public partial class DataForm : Window
     {
         public DataForm()
@@ -26,36 +23,33 @@ namespace HospitalEmergencySimulation
         }
         private void TextBox_PreviewTextInput(object sender, TextCompositionEventArgs e)
         {
-            // Comprobar si el texto ingresado es numérico
             if (!IsTextNumeric(e.Text))
             {
-                // Si no es numérico, cancelar el evento
                 e.Handled = true;
             }
         }
         private bool IsTextNumeric(string text)
         {
-            // Intentar convertir el texto a un número
             return double.TryParse(text, out _);
         }
         private void getParameters(object sender, RoutedEventArgs e)
         {
-            
-            int vMinimumAttentionTimeHighPriority = int.Parse(minimumAttentionTimeHighPriority.Text);
-            int vMaximumAttentionTimeHighPriority = int.Parse(maximumAttentionTimeHighPriority.Text);
-            int vMinimumAttentionTimeLowPriority = int.Parse(minimumAttentionTimeLowPriority.Text);
-            int vMaximumAttentionTimeLowPriority = int.Parse(maximumAttentionTimeLowPriority.Text);
-            int vLambdaArrivalHighPrority = int.Parse(lambdaArrivalHighPrority.Text);
-            int vLambdaArrivalLowPrority = int.Parse(lambdaArrivalLowPrority.Text);
+
+            double vMinimumAttentionTimeHighPriority = double.Parse(minimumAttentionTimeHighPriority.Text);
+            double vMaximumAttentionTimeHighPriority = double.Parse(maximumAttentionTimeHighPriority.Text);
+            double vMinimumAttentionTimeLowPriority = double.Parse(minimumAttentionTimeLowPriority.Text);
+            double vMaximumAttentionTimeLowPriority = double.Parse(maximumAttentionTimeLowPriority.Text);
+            double vLambdaArrivalHighPrority = double.Parse(lambdaArrivalHighPrority.Text);
+            double vLambdaArrivalLowPrority = double.Parse(lambdaArrivalLowPrority.Text);
             int vNumberArrivalIntervals = int.Parse(numberArrivalIntervals.Text);
 
             ControllerSimulation controller = new ControllerSimulation(vMinimumAttentionTimeHighPriority, vMaximumAttentionTimeHighPriority, vMinimumAttentionTimeLowPriority, vMaximumAttentionTimeLowPriority,
-                vLambdaArrivalHighPrority, vLambdaArrivalLowPrority, vNumberArrivalIntervals);
+            vLambdaArrivalHighPrority, vLambdaArrivalLowPrority, vNumberArrivalIntervals);
             MainWindow mainWindow = new MainWindow(controller);
             mainWindow.Show();
             this.Close();
         }
-        
+
     }
 
 }
